+++ android/src/main/java/io/github/zileyuan/umeng_analytics_push/SimpleMeizuPushReceiver.java
@@ -1,0 +1,6 @@
+package io.github.zileyuan.umeng_analytics_push;
+
+import org.android.agoo.mezu.MeizuPushReceiver;
+
+public class SimpleMeizuPushReceiver extends MeizuPushReceiver {
+}