+++ android/src/main/java/io/github/zileyuan/umeng_analytics_push/UmengAnalyticsPushFlutterAndroid.java
@@ -1,14 +1,21 @@
 package io.github.zileyuan.umeng_analytics_push;
 
+import android.app.Application;
+import android.app.Activity;
 import android.util.Log;
 import android.content.Context;
 import com.umeng.commonsdk.UMConfigure;
 import com.umeng.message.IUmengRegisterCallback;
 import com.umeng.analytics.MobclickAgent;
 import com.umeng.message.PushAgent;
+import com.umeng.message.MsgConstant;
 import com.umeng.message.UmengNotificationClickHandler;
 import com.umeng.message.entity.UMessage;
 
+import org.android.agoo.huawei.HuaWeiRegister;
+import org.android.agoo.mezu.MeizuRegister;
+import org.android.agoo.oppo.OppoRegister;
+import org.android.agoo.vivo.VivoRegister;
 
 public class UmengAnalyticsPushFlutterAndroid {
 
@@ -21,6 +28,14 @@ public class UmengAnalyticsPushFlutterAndroid {
         if (!messageSecret.isEmpty()) {
             //获取消息推送代理示例
             PushAgent mPushAgent = PushAgent.getInstance(context);
+            //设置客户端允许声音提醒
+            mPushAgent.setNotificationPlaySound(MsgConstant.NOTIFICATION_PLAY_SDK_ENABLE);
+            //设置客户端允许呼吸灯点亮
+            mPushAgent.setNotificationPlayLights(MsgConstant.NOTIFICATION_PLAY_SDK_ENABLE);
+            //设置客户端允许震动
+            mPushAgent.setNotificationPlayVibrate(MsgConstant.NOTIFICATION_PLAY_SDK_ENABLE);
+            //设置冷却时间，避免三分钟内出现多条通知而被替换
+            mPushAgent.setMuteDurationSeconds(180);
             //注册推送服务，每次调用register方法都会回调该接口
             mPushAgent.register(new IUmengRegisterCallback() {
                 @Override
@@ -56,5 +71,25 @@ public class UmengAnalyticsPushFlutterAndroid {
 
     public static void androidOnPause(Context context) {
         MobclickAgent.onPause(context);
+    }
+
+    public static void registerXiaomi(Context context, String appId, String appKey) {
+        MiPushRegistar.register(context, appId, appKey);
+    }
+
+    public static void registerHuawei(Application application) {
+        HuaWeiRegister.register(application);
+    }
+
+    public static void registerOppo(Context context, String appKey, String appSecret) {
+        OppoRegister.register(context, appKey, appSecret);
+    }
+
+    public static void registerVivo(Context context) {
+        VivoRegister.register(context);
+    }
+
+    public static void registerMeizu(Context context, String appId, String appKey) {
+        MeizuRegister.register(context, appId, appKey);
     }
 }