+++ ios/Classes/UmengAnalyticsPushIos.m
@@ -1,6 +1,6 @@
 #import "UmengAnalyticsPushIos.h"
 #import <UMCommon/UMCommon.h>
-#import <UMAnalytics/MobClick.h>
+#import <UMCommon/MobClick.h>
 #import <UMPush/UMessage.h>
 #import <UserNotifications/UserNotifications.h>
 #import <UMCommonLog/UMCommonLogHeaders.h>